requirements.txt
discord.py==1.4.1
praw==7.1.0
